+++ capture-search-result/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to the "capture-search-result" extension will be documented in this file.
 
+## [0.0.4]
+
+- Include column title row in output.  
+- Add two settings.  
+
+    1. Include search summary when copying search results. True: include / False: do not include
+    1. Include column title row when copying search results. True: include / False: do not include
+
 ## [0.0.3]
 
 Include search result summary in output.